CMS RuDi 0.0.7 (Instant CMS v1.10.4 DS Soft Mod v 0.0.2) -------------------------------------------------------------------------------- [+] Добавлена возможность настраивать плагины в отключенном состоянии; [+] В класс плагинов cmsPlugin добавлен метод удаления плагина uninstall, который     могут использовать разработчики в своих плагинах для отлавливания момента     удаления плагина; [=] Капча в системе теперь подключается с помощью плагинов с типом captcha.     Для того чтобы разобраться как писать свои плагины для вывода капчи можете     посмотреть как устроен системный плагин p_captcha; [+] В систему добавлен плагин p_recaptcha позволяющий использовать капчу от гугла     https://www.google.com/recaptcha/; [=] Исправлены ошибки: с выдачей неизвестной ошибки при мультизагрузке фотографий     к статье; в php шаблоне плагина p_content_imgs; [+] В систему добавлен Bootstrap v3.2.0 (Иконки Font Awesome v4.1.0) в папку     /includes/bootstrap/ можете подключать его в своих разработках, в админке     Bootstrap подключен по умолчанию; [-] Удален файл /includes/tools.inc.php; [=] Изменен дизайн админки. Возможны небольщие баги в виде админки но функционал     полностью рабочий.; [=] Исправлены несколько ошибок в классе curl.class.php; [+] Добавлено определение работы сайта по https или http и соответственно     присваивание правильного значения для константы HOST; [+] Добавлена поддержка работы системы как обычного сайта, так и iframe приложения     социальных сетей. Пока поддерживается только "Вконтакте"; [+] Добавлена поддержка поискового движка Sphinx; [+] В настройки сайта добавлена возможность смены компонента "по умолчанию". У     компонента "по умолчанию" в ссылке можно не указывать название компонента     как например у компонента "Каталог статей" не указывается /content/; [+] Для шаблонов добавлена возможность иметь свои настройки, для этого в папке     шаблона нужно создать файл config.php в которой функция get_template_cfg_fields     должна возвращать массив полей с настройками, а функция get_template_default_cfg     должна возвращать массив с дефолтовыми настройками. Настроить параметры шаблона     можно в Админка->Настройки сайта->Дизайн->Настройки шаблона; [+] Добавлено объединение всех js и css файлов в один файл js и css. Опция тестовая,     дальнейшая ее судьба не определена. В данный момент она работает, но единый     файл css и js создается при каждой загрузке страницы из за чего он не кешируется     браузером; [+] Для плагина p_captcha добавлена поддержка нескольких капч на одной странице.     Внимание. Для плагина p_recaptcha невозможно использовать несколько капч     на одной странице (если у кого то есть информация о том как такое реализовать     напишите в поддержку добавлю и для него) поэтому не рекомендуется использовать     несколько капч на одной странице; [+] Добавлен еще один плагин для вывода каптчи p_phpcaptcha с сайта https://www.phpcaptcha.org     пока нет поддержки мультикапч на одной странице, это будет реализовано     в следующем обновлении системы. Но те кто не использует несколько капч на     одной странице уже могут использовать этот плагин; [=] Подправлен вывод капчи плагином p_recaptcha для комментариев;
Внесены изменения http://trac.instantcms.ru/changeset/1416 от fuze (06/26/14 16:08:35);
Внесены изменения http://trac.instantcms.ru/changeset/1417 от fuze (07/02/14 12:43:20);
Внесены изменения http://trac.instantcms.ru/changeset/1418 от fuze (07/05/14 21:51:33);
Внесены изменения http://trac.instantcms.ru/changeset/1419 от fuze (07/05/14 23:06:09);
Внесены изменения http://trac.instantcms.ru/changeset/1420 от fuze (07/13/14 22:50:04);
        разное по безопасности +
        возможность для модулей исключать страницы для показа: ставить чек-бокс "Показывать на всех страницах сайта" и выбирать пункты меню, на которых модуль не показывать

!Не Внесены изменения http://trac.instantcms.ru/changeset/1422 от maxisoft (07/14/14 21:16:25);
    Причина: Аналогичные изменения в CMS RuDi сделаны ранее.

Внесены изменения http://trac.instantcms.ru/changeset/1424 от fuze (07/16/14 16:27:55);
Внесены изменения http://trac.instantcms.ru/changeset/1425 от maxisoft (07/24/14 11:17:39);
        Добавлен тип HTML (поле) в настройках модуля.
Внесены изменения http://trac.instantcms.ru/changeset/1427 от fuze (07/26/14 19:03:52);
Внесены изменения http://trac.instantcms.ru/changeset/1429 от fuze (07/31/14 21:58:18);
        поиск с изображениями
Внесены изменения http://trac.instantcms.ru/changeset/1431 от maxisoft (08/24/14 10:21:23);
        Вынесен в файл шаблона вывод pathway
Внесены изменения http://trac.instantcms.ru/changeset/1433 от maxisoft (09/15/14 22:43:52);
        Обновлена версия Smarty до 3.1.19
Внесены изменения http://trac.instantcms.ru/changeset/1434 от maxisoft (09/15/14 22:45:28);
        Вынесен pagebar в отдельный шаблон.

!Не Внесены изменения http://trac.instantcms.ru/changeset/1435 от maxisoft (09/15/14 22:53:59);
    Причина: Эти и некоторые другие аналогичные изменения были выполнены после обновления Smarty

[+] В класс phpTpl добавлен новый метод fetch() который возвращает вывод шаблона
    вместо его отображения на экран. В smarty этот метод был доступен всегда;
[+] Добавлена сортировка списков генерируемых функцией cpListTable, при клике
    по заголовку поля загружается список с сортировкой по выбранному параметру, при
    повторном клике по тому же заголовку загружается список с обратной сортировкой;
[=] Рефакторинг кода класса smartyTpl. При выводе шаблонов, методы display и fetch
    нужно вызывать без параметров просто ->display() и ->fetch(). Так же название
    файла шаблона нужно передавать без расширения
    Пример:
        Не правильно
            self::initTemplate('modules', 'module.tpl')->
                        ...
                        display('module.tpl');
        Правильно
            self::initTemplate('modules', 'module')->
                        ...
                        display();
[+] В систему добавлен Тикет раздел где можно написать в тех поддержку о найденных
    ошибках в системе или предложить идею по новому функционалу.
================================================================================

+++ admin/applets/plugins.php
@@ -13,92 +13,82 @@
 
 if(!defined('VALID_CMS_ADMIN')) { die('ACCESS DENIED'); }
 
-function applet_plugins(){
-
+function applet_plugins() {
     global $_LANG;
-
+    
     $inCore = cmsCore::getInstance();
 
     cmsCore::c('page')->setAdminTitle($_LANG['AD_PLUGINS']);
     cpAddPathway($_LANG['AD_PLUGINS'], 'index.php?view=plugins');
+    
+    global $adminAccess;
+    if (!cmsUser::isAdminCan('admin/plugins', $adminAccess)) { cpAccessDenied(); }
 
-	global $adminAccess;
-	if (!cmsUser::isAdminCan('admin/plugins', $adminAccess)) { cpAccessDenied(); }
+    $do = cmsCore::request('do', 'str', 'list');
+    $id = cmsCore::request('id', 'int', -1);
 
-	$do = cmsCore::request('do', 'str', 'list');
-	$id = cmsCore::request('id', 'int', -1);
+    if ($do == 'hide') {
+        cmsCore::c('db')->setFlag('cms_plugins', $id, 'published', '0');
+        cmsCore::halt('1');
+    }
 
-// ===================================================================================== //
+    if ($do == 'show') {
+        cmsCore::c('db')->setFlag('cms_plugins', $id, 'published', '1');
+        cmsCore::halt('1');
+    }
 
-	if ($do == 'hide'){
-            cmsCore::c('db')->setFlag('cms_plugins', $id, 'published', '0');
-            cmsCore::halt('1');
-	}
+    if ($do == 'list') {
+        $toolmenu = array(
+            array( 'icon' => 'install.gif', 'title' => $_LANG['AD_INSTALL_PLUGINS'], 'link' => '?view=install&do=plugin' )
+        );
 
-// ===================================================================================== //
-
-	if ($do == 'show'){
-            cmsCore::c('db')->setFlag('cms_plugins', $id, 'published', '1');
-            cmsCore::halt('1');
-	}
-
-// ===================================================================================== //
-
-	if ($do == 'list'){
-
-		$toolmenu = array();
-		$toolmenu[1]['icon']    = 'install.gif';
-		$toolmenu[1]['title']   = $_LANG['AD_INSTALL_PLUGINS'];
-		$toolmenu[1]['link']    = '?view=install&do=plugin';
-
-		cpToolMenu($toolmenu);
+        cpToolMenu($toolmenu);
 
         $plugin_id = cmsCore::request('installed', 'str', '');
 
-        if ($plugin_id){
+        if ($plugin_id) {
             $task = cmsCore::request('task', 'str', 'install');
 
-            if ($task == 'install' || $task == 'upgrade'){
-                $plugin     = $inCore->loadPlugin($plugin_id);
-                $task_str   = ($task=='install') ? $_LANG['AD_IS_INSTALL'] : $_LANG['AD_IS_UPDATE'];
-                echo '<div style="color:green;margin-top:12px;margin-bottom:5px;">'.$_LANG['AD_PLUGIN'].' <strong>"'.$plugin->info['title'].'"</strong> '.$task_str.'. '.$_LANG['AD_ENABLE_PLUGIN'].'.</div>';
+            if ($task == 'install' || $task == 'upgrade') {
+                $plugin   = $inCore->loadPlugin($plugin_id);
+                $task_str = ($task=='install') ? $_LANG['AD_IS_INSTALL'] : $_LANG['AD_IS_UPDATE'];
+                echo '<div style="color:green;margin-top:12px;margin-bottom:5px;">'. $_LANG['AD_PLUGIN'] .' <strong>"'. $plugin->info['title'] .'"</strong> '. $task_str .'. '. $_LANG['AD_ENABLE_PLUGIN'] .'.</div>';
             }
 
-            if ($task == 'remove'){
-                echo '<div style="color:green;margin-top:12px;margin-bottom:5px;">'.$_LANG['AD_REMOVE_PLUGIN_OK'].'.</div>';
+            if ($task == 'remove') {
+                echo '<div style="color:green;margin-top:12px;margin-bottom:5px;">'. $_LANG['AD_REMOVE_PLUGIN_OK'] .'.</div>';
             }
         }
 
-		$fields = array();
+        $fields = array(
+            array( 'title' => 'id', 'field' => 'id', 'width' => '40' ),
+            array( 'title' => $_LANG['TITLE'], 'field' => 'title', 'width' => '250' ),
+            array( 'title' => $_LANG['DESCRIPTION'], 'field' => 'description', 'width' => '' ),
+            array( 'title' => $_LANG['AD_AUTHOR'], 'field' => 'author', 'width' => '160' ),
+            array( 'title' => $_LANG['AD_VERSION'], 'field' => 'version', 'width' => '80' ),
+            array( 'title' => $_LANG['AD_FOLDER'], 'field' => 'plugin', 'width' => '100' ),
+            array( 'title' => $_LANG['AD_ENABLE'], 'field' => 'published', 'width' => '80' ),
+        );
 
-		$fields[0]['title'] = 'id';			$fields[0]['field'] = 'id';			$fields[0]['width'] = '20';
-		$fields[1]['title'] = $_LANG['TITLE'];       $fields[1]['field'] = 'title';		  $fields[1]['width'] = '250';
-		$fields[2]['title'] = $_LANG['DESCRIPTION']; $fields[2]['field'] = 'description'; $fields[2]['width'] = '';
-		$fields[3]['title'] = $_LANG['AD_AUTHOR'];   $fields[3]['field'] = 'author';	  $fields[3]['width'] = '160';
-		$fields[4]['title'] = $_LANG['AD_VERSION'];  $fields[4]['field'] = 'version';     $fields[4]['width'] = '50';
-        $fields[5]['title'] = $_LANG['AD_FOLDER'];   $fields[5]['field'] = 'plugin';      $fields[5]['width'] = '100';
-		$fields[6]['title'] = $_LANG['AD_ENABLE'];	 $fields[6]['field'] = 'published';	  $fields[6]['width'] = '60';
+        $actions = array(
+            array( 'title' => $_LANG['AD_CONFIG'], 'icon' => 'config.gif', 'link' => '?view=plugins&do=config&id=%id%' ),
+            array( 'title' => $_LANG['DELETE'], 'icon' => 'delete.gif', 'link' => '?view=install&do=remove_plugin&id=%id%', 'confirm' => $_LANG['AD_REMOVE_PLUGIN_FROM'] ),
+        );
 
-		$actions = array();
-		$actions[0]['title'] = $_LANG['AD_CONFIG'];
-		$actions[0]['icon']  = 'config.gif';
-		$actions[0]['link']  = '?view=plugins&do=config&id=%id%';
+        cpListTable('cms_plugins', $fields, $actions);
+    }
 
-		$actions[1]['title']   = $_LANG['DELETE'];
-		$actions[1]['icon']    = 'delete.gif';
-		$actions[1]['confirm'] = $_LANG['AD_REMOVE_PLUGIN_FROM'];
-		$actions[1]['link']    = '?view=install&do=remove_plugin&id=%id%';
-
-		cpListTable('cms_plugins', $fields, $actions);
-
-	}
-
-// ===================================================================================== //
-
-    if ($do == 'save_config'){
+    if ($do == 'save_config') {
         if (!cmsUser::checkCsrfToken()) { cmsCore::error404(); }
-        $plugin_name    = cmsCore::request('plugin', 'str', 0);
-        $config         = cmsCore::request('config', 'array_str');
+        
+        $plugin_name = cmsCore::request('plugin', 'str', 0);
+        $plugin = $inCore->loadPlugin($plugin_name);
+        
+        if (!method_exists($plugin, 'getConfigFields')) {
+            $config = cmsCore::request('config', 'array_str');
+        } else {
+            $config = cmsCore::c('form_gen')->requestForm($plugin->getConfigFields());
+        }
 
         if (!$config || !$plugin_name) { cmsCore::redirectBack(); }
 
@@ -106,57 +96,49 @@ function applet_plugins(){
 
         cmsUser::clearCsrfToken();
 
-		cmsCore::addSessionMessage($_LANG['AD_CONFIG_SAVE_SUCCESS'], 'success');
-
+        cmsCore::addSessionMessage($_LANG['AD_CONFIG_SAVE_SUCCESS'], 'success');
         cmsCore::redirect('index.php?view=plugins');
-
     }
 
-// ===================================================================================== //
-
-    if ($do == 'config'){
-
+    if ($do == 'config') {
         $plugin_name = $inCore->getPluginById($id);
-        if(!$plugin_name){ cmsCore::error404(); }
+        if (!$plugin_name) { cmsCore::error404(); }
 
         $plugin = $inCore->loadPlugin($plugin_name);
         $config = $inCore->loadPluginConfig($plugin_name);
 
         cmsCore::c('page')->setAdminTitle($plugin->info['title']);
-        cpAddPathway($plugin->info['title'], 'index.php?view=plugins&do=config&id='.$id);
+        cpAddPathway($plugin->info['title'], 'index.php?view=plugins&do=config&id='. $id);
 
-        echo '<h3>'.$plugin->info['title'].'</h3>';
+        echo '<h3>'. $plugin->info['title'] .'</h3>';
 
         if (!$config) {
-            echo '<p>'.$_LANG['AD_PLUGIN_DISABLE'].'.</p>';
-            echo '<p><a href="javascript:window.history.go(-1);">'.$_LANG['BACK'].'</a></p>';
+            echo '<p>'. $_LANG['AD_PLUGIN_DISABLE'] .'.</p>';
+            echo '<p><a href="javascript:window.history.go(-1);">'. $_LANG['BACK'] .'</a></p>';
             return;
 
         }
 
-        echo '<form action="index.php?view=plugins&do=save_config&plugin='.$plugin_name.'" method="POST">';
-            echo '<input type="hidden" name="csrf_token" value="'.cmsUser::getCsrfToken().'" />';
-
-            echo '<table class="proptable" width="605" cellpadding="8" cellspacing="0" border="0">';
-                foreach ($config as $field=>$value){
-                    echo '<tr>';
-                        echo '<td width="150"><strong>'.(isset($_LANG[mb_strtoupper($field)]) ? $_LANG[mb_strtoupper($field)] : $field).':</strong></td>';
-                        echo '<td><input type="text" style="width:90%" name="config['.$field.']" value="'.htmlspecialchars($value).'" /></td>';
-                    echo '</tr>';
-                }
-            echo '</table>';
+        echo '<form action="index.php?view=plugins&do=save_config&plugin='. $plugin_name .'" method="POST">';
+            if (!method_exists($plugin, 'getConfigFields')) {
+                echo '<input type="hidden" name="csrf_token" value="'. cmsUser::getCsrfToken() .'" />';
+                echo '<table class="proptable" width="605" cellpadding="8" cellspacing="0" border="0">';
+                    foreach ($config as $field => $value) {
+                        echo '<tr>';
+                            echo '<td width="150"><strong>'. cmsCore::getArrVal($_LANG, mb_strtoupper($field), $field) .':</strong></td>';
+                            echo '<td><input type="text" style="width:90%" name="config['. $field .']" value="'. htmlspecialchars($value) .'" /></td>';
+                        echo '</tr>';
+                    }
+                echo '</table>';
+            } else {
+                echo '<div style="width:650px;">'. cmsCore::c('form_gen')->generateForm($plugin->getConfigFields(), $config) .'</div>';
+            }
 
             echo '<div style="margin-top:6px;">';
-                echo '<input type="submit" name="save" value="'.$_LANG['SAVE'].'" /> ';
-                echo '<input type="button" name="back" value="'.$_LANG['CANCEL'].'" onclick="window.history.go(-1)" />';
+                echo '<input type="submit" class="btn btn-primary" name="save" value="'. $_LANG['SAVE'] .'" /> ';
+                echo '<input type="button" class="btn btn-default" name="back" value="'. $_LANG['CANCEL'] .'" onclick="window.history.go(-1)" />';
             echo '</div>';
 
         echo '</form>';
-
     }
-
-// ===================================================================================== //
-
 }
-
-?>
